move niftyreg
