+++ brainles_preprocessing/registration/__init__.py
@@ -1,1 +1,1 @@
-from .niftyreg import NiftyRegRegistrator
+from .niftyreg.niftyreg import NiftyRegRegistrator
